Move BasePrice grouping method into BasePrice class
