+++ src/main/java/com/applicant/redbubble/cart_calculator/models/BasePrice.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 public class BasePrice {
     @JsonProperty("product-type")
@@ -25,5 +26,13 @@ public class BasePrice {
 
     public int getBasePrice() {
         return basePrice;
+    }
+
+    public static Map<String, List<BasePrice>> groupPricesByProductType(List<BasePrice> basePrices) {
+        if (basePrices != null && basePrices.size() != 0) {
+            return basePrices.stream().collect(Collectors.groupingBy(BasePrice::getProductType));
+        } else {
+            return null;
+        }
     }
 }
